+++ src/org-manager/org-manager.routing.ts
@@ -4,7 +4,7 @@ import { ActiveOrganisationsComponent } from './containers';
 import { PendingOrganisationsComponent} from 'src/org-manager/containers/pending-organisations/pending-organisations.component';
 import { ApproveOrganisationComponent } from 'src/org-manager/containers/approve-organisation/approve-organisation.component';
 import { ApproveOrganisationSuccessComponent } from 'src/org-manager/containers/approve-organisation-success/approve-organisation-success.component';
-import { AuthGuard } from 'src/services/auth/auth.guard';
+import { AuthGuard } from 'src/app/services/auth/auth.guard';
 import { OrganisationDetailsComponent } from './components';
 import { EditDetailsComponent } from './containers/edit-details/edit-details.component';
